# services/authorization-service/src/authz_service/api/generated/routers/protocols.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-09-22T10:20:20+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=["protocols"])


@router.post(
    "/v1/decisions/protocols/create",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["protocols"],
)
def authorize_protocol_create(body: ProtocolCreateRequest) -> Union[Decision, Problem]:
    """
    May the subject add a protocol to this RSE?
    """
    pass


@router.post(
    "/v1/decisions/protocols/delete",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["protocols"],
)
def authorize_protocol_delete(body: ProtocolDeleteRequest) -> Union[Decision, Problem]:
    """
    May the subject delete a protocol from this RSE?
    """
    pass


@router.post(
    "/v1/decisions/protocols/update",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["protocols"],
)
def authorize_protocol_update(body: ProtocolUpdateRequest) -> Union[Decision, Problem]:
    """
    May the subject update a protocol on this RSE?
    """
    pass