# services/authorization-service/src/authz_service/api/generated/routers/rses.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-09-22T10:20:20+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=["rses"])


@router.post(
    "/v1/decisions/rses/attributes/delete",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rses"],
)
def authorize_rse_attribute_delete(
    body: RseAttributeDeleteRequest,
) -> Union[Decision, Problem]:
    """
    May the subject delete an attribute from this RSE?
    """
    pass


@router.post(
    "/v1/decisions/rses/attributes/set",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rses"],
)
def authorize_rse_attribute_set(
    body: RseAttributeSetRequest,
) -> Union[Decision, Problem]:
    """
    May the subject set an attribute on this RSE?
    """
    pass


@router.post(
    "/v1/decisions/rses/create",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rses"],
)
def authorize_rse_create(body: RseCreateRequest) -> Union[Decision, Problem]:
    """
    May the subject create an RSE with this name?
    """
    pass


@router.post(
    "/v1/decisions/rses/delete",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rses"],
)
def authorize_rse_delete(body: RseDeleteRequest) -> Union[Decision, Problem]:
    """
    May the subject delete this RSE?
    """
    pass


@router.post(
    "/v1/decisions/rses/update",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rses"],
)
def authorize_rse_update(body: RseUpdateRequest) -> Union[Decision, Problem]:
    """
    May the subject update this RSE, including a rename?
    """
    pass