# services/authorization-service/src/authz_service/api/generated/routers/rules.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-09-22T10:20:20+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=["rules"])


@router.post(
    "/v1/decisions/rules/create",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rules"],
)
def authorize_rule_create(body: RuleCreateRequest) -> Union[Decision, Problem]:
    """
    May the subject create a replication rule over these DIDs?
    """
    pass


@router.post(
    "/v1/decisions/rules/delete",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rules"],
)
def authorize_rule_delete(body: RuleDeleteRequest) -> Union[Decision, Problem]:
    """
    May the subject delete an existing rule?
    """
    pass


@router.post(
    "/v1/decisions/rules/update",
    response_model=Decision,
    responses={
        "400": {"model": Problem},
        "401": {"model": Problem},
        "403": {"model": Problem},
        "500": {"model": Problem},
    },
    tags=["rules"],
)
def authorize_rule_update(body: RuleUpdateRequest) -> Union[Decision, Problem]:
    """
    May the subject change an existing rule?
    """
    pass